remove local signature agg dependency / refactoring
