+++ components/tools/common/utils/validation.ts
@@ -24,3 +24,16 @@ export const isValidAllowlistPrecompileConfig = (config: AllowlistPrecompileConf
   return !Object.values(config.addresses).some(entries => hasErrors(entries as AddressEntry[]));
 }
 
+export function isValidUrl(url: string): boolean {
+    try {
+        new URL(url);
+        return true;
+    } catch {
+        return false;
+    }
+}
+
+export function isValidAddress(address: string): boolean {
+    return /^0x[a-fA-F0-9]{40}$/.test(address);
+}
+
